more modal stuff

+++ Data/Models/Product.cs
@@ -4,7 +4,7 @@ public record Product
 {
     public required int Model { get; set; }
     public required string Name { get; set; }
-    public required int Price { get; set; }
+    public required decimal Price { get; set; }
     public required string Description { get; set; }
     public required string Image { get; set; }
 }
